+++ tiger-rbel/src/main/java/de/gematik/rbellogger/converter/RbelBase64JsonConverter.java
@@ -17,39 +17,40 @@ public class RbelBase64JsonConverter extends RbelJsonConverter {
 
   @Override
   public void consumeElement(RbelElement rbel, RbelConverter context) {
-    if (rbel.getRawStringContent().isEmpty()) {
+    if (rbel.getRawContent().length == 0) {
       return;
     }
-    safeConvertBase64Using(rbel.getRawStringContent(), Base64.getDecoder(), context, rbel)
+    safeConvertBase64Using(rbel.getRawContent(), Base64.getDecoder(), context, rbel)
         .or(
             () ->
                 safeConvertBase64Using(
-                    rbel.getRawStringContent(), Base64.getUrlDecoder(), context, rbel))
+                    rbel.getRawContent(), Base64.getUrlDecoder(), context, rbel))
         .ifPresent(rbel::addFacet);
   }
 
   private Optional<RbelBase64Facet> safeConvertBase64Using(
-      String input, Decoder decoder, RbelConverter context, RbelElement parentNode) {
-    return Optional.ofNullable(input)
-        .map(
-            i -> {
-              try {
-                return decoder.decode(i);
-              } catch (IllegalArgumentException e) {
-                return null;
-              }
-            })
-        .map(data -> new RbelElement(data, parentNode))
-        .stream()
-        .peek(context::convertElement)
-        .filter(innerNode -> innerNode.hasFacet(RbelRootFacet.class))
-        .filter(
-            innerNode ->
-                innerNode.getFacetOrFail(RbelRootFacet.class).getRootFacet()
-                        instanceof RbelJsonFacet
-                    || innerNode.getFacetOrFail(RbelRootFacet.class).getRootFacet()
-                        instanceof RbelXmlFacet)
-        .map(RbelBase64Facet::new)
-        .findAny();
+      byte[] input, Decoder decoder, RbelConverter context, RbelElement parentNode) {
+    final Optional<RbelElement> rawElement = Optional.ofNullable(input)
+      .map(
+        i -> {
+          try {
+            return decoder.decode(i);
+          } catch (IllegalArgumentException e) {
+            return null;
+          }
+        })
+      .map(data -> new RbelElement(data, parentNode));
+    if (rawElement.isEmpty()) {
+      return Optional.empty();
+    }
+    context.convertElement(rawElement.get());
+    if (rawElement.get().hasFacet(RbelRootFacet.class)) {
+      final RbelRootFacet<?> rootFacet = rawElement.get().getFacetOrFail(RbelRootFacet.class);
+      if (rootFacet.getRootFacet() instanceof RbelJsonFacet
+          || rootFacet.getRootFacet() instanceof RbelXmlFacet) {
+        return Optional.of(new RbelBase64Facet(rawElement.get()));
+      }
+    }
+    return Optional.empty();
   }
 }