+++ tiger-rbel/src/main/java/de/gematik/rbellogger/data/facet/RbelPrintableHashFacet.java
@@ -1,0 +1,51 @@
+/*
+ * ${GEMATIK_COPYRIGHT_STATEMENT}
+ */
+
+package de.gematik.rbellogger.data.facet;
+
+import static de.gematik.rbellogger.renderer.RbelHtmlRenderingToolkit.*;
+import static j2html.TagCreator.div;
+import static j2html.TagCreator.h3;
+
+import de.gematik.rbellogger.data.RbelElement;
+import de.gematik.rbellogger.data.RbelMultiMap;
+import de.gematik.rbellogger.renderer.RbelHtmlFacetRenderer;
+import de.gematik.rbellogger.renderer.RbelHtmlRenderer;
+import de.gematik.rbellogger.renderer.RbelHtmlRenderingToolkit;
+import j2html.tags.ContainerTag;
+import java.util.Optional;
+import lombok.Value;
+
+@Value
+public class RbelPrintableHashFacet implements RbelFacet {
+
+  static {
+    RbelHtmlRenderer.registerFacetRenderer(
+        new RbelHtmlFacetRenderer() {
+          @Override
+          public boolean checkForRendering(RbelElement element) {
+            return element.hasFacet(RbelPrintableHashFacet.class);
+          }
+
+          @Override
+          public ContainerTag performRendering(
+              RbelElement element,
+              Optional<String> key,
+              RbelHtmlRenderingToolkit renderingToolkit) {
+            final var hashFacet = element.getFacet(RbelPrintableHashFacet.class).orElseThrow();
+            return div()
+                .with(h3(hashFacet.getTitle()))
+                .with(renderingToolkit.printAsBinary(element))
+                .with(addNotes(element));
+          }
+        });
+  }
+
+  String title;
+
+  @Override
+  public RbelMultiMap<RbelElement> getChildElements() {
+    return new RbelMultiMap<>();
+  }
+}